Rename `ws` to `websocket`
